app/requirements.txt
django~=3.2.16
python-dotenv~=0.21.0
django-split-settings~=1.2.0
django-debug-toolbar~=3.4.0
psycopg2-binary~=2.9
asgiref~=3.5.2
pytz~=2022.6
sqlparse~=0.4.3
uwsgi~=2.0.20
django-extensions~=3.2.1
django-cors-headers~=3.13.0
djangorestframework~=3.14.0